+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project are documented in this file.
    color scheme is dark.
  - Added `fluent-light` and `fluent-dark` as explicit styles to select a light/dark variant,
    regardless of the system color scheme setting.
+ - Added `Window::is_visible` 
 
 ### Fixed